+++ SolidModelBrowser/HelpString.cs
@@ -29,6 +29,7 @@ O - Show/hide XYZ axes
 P - Show/hide ground polygon
 CTRL+F - Recreate smooth mesh in flat mode (flat polygon lighting)
 CTRL+W - Wireframe mode
+CTRL+S - Export current model
 I - show/hide model info
 LeftShift - hold left shift to drag window by viewport